pyproject.toml
[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "personal-brand-autoposter"
version = "0.1.0"
description = "Personalized X/Twitter voice content generator, scheduler, and autoposter."
readme = "README.md"
requires-python = ">=3.10"
authors = [
  {name = "Chudi Nnorukam"}
]
keywords = ["twitter", "automation", "scheduler", "content-generation"]
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  "typer>=0.12,<0.13",
  "rich>=13,<14",
  "shellingham>=1.3,<2",
  "click>=8.1.7,<8.2",
  "pydantic>=1.10,<2.0",
  "python-dotenv>=1.0,<2.0",
  "tweepy>=4.14,<5.0",
  "snscrape==0.7.0.20230622",
  "APScheduler>=3.10,<4.0",
  "openai>=1.3,<2.0",
  "httpx>=0.23,<0.25",
  "requests>=2.31,<3.0",
  "pandas>=2.1,<3.0",
]

[project.scripts]
x-autoposter = "autoposter.cli:app"

[project.optional-dependencies]
develop = [
  "ruff",
  "pytest",
]

[tool.setuptools.packages.find]
include = ["autoposter*"]
exclude = ["data*"]
